+++ src/extension/server/tsnode.ts
@@ -4,6 +4,6 @@ import * as path from 'path';
 const tsNodePath = path.join(__dirname, '..', '..', '..', 'resources', 'scripts', 'node_modules', 'ts-node');
 export function register(context: vm.Context) {
     vm.runInNewContext(`require('${tsNodePath.replace(/\\/g, '/')}').register()`, context, {
-        displayErrors: true
+        displayErrors: false
     });
 }