+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.3 (10 Sept 2021)
 * Update samples to use `isomorphic-fetch` instead of `node-fetch` (and pre-requisite `npm` packages).
+* Basic support for user input in notebooks using [readline](https://nodejs.org/api/readline.html#readline_readline_createinterface_options)
 
 ## 2.0.2 (6 Sept 2021)
 * Excellent support for [arquero](https://uwdata.github.io/arquero/) (rich HTML output)